+++ playground/next.config.ts
@@ -1,0 +1,32 @@
+import type { NextConfig } from "next";
+
+const nextConfig: NextConfig = {
+  cacheComponents: true,
+  logging: {
+    fetches: {
+      fullUrl: true,
+    },
+  },
+  images: {
+    remotePatterns: [
+      {
+        protocol: "https",
+        hostname: "lab.basement.studio",
+      },
+      {
+        protocol: "https",
+        hostname: "assets.basehub.com",
+      },
+      {
+        protocol: "https",
+        hostname: "basehub.earth",
+      },
+    ],
+  },
+};
+
+const withBundleAnalyzer = require("@next/bundle-analyzer")({
+  enabled: process.env.ANALYZE === "true",
+});
+
+export default withBundleAnalyzer(nextConfig);