+++ playground/src/app/page.tsx
@@ -1,39 +1,17 @@
-import { Pump } from "basehub/react-pump";
+import Link from "next/link";
 
 export default async function HomePage() {
   return (
-    <>
-      <Pump
-        queries={[
-          {
-            blog: {
-              __args: {
-                variants: {
-                  language: "en",
-                },
-              },
-              _title: true,
-              posts: {
-                item: {
-                  _id: true,
-                },
-              },
-            },
-          },
-        ]}
-      >
-        {async ([data]) => {
-          "use server";
-
-          return (
-            <div>
-              <pre>
-                <code>{JSON.stringify(data, null, 2)}</code>
-              </pre>
-            </div>
-          );
-        }}
-      </Pump>
-    </>
+    <div
+      style={{
+        display: "flex",
+        flexDirection: "column",
+        gap: 12,
+        fontFamily: "sans-serif",
+      }}
+    >
+      <Link href="/client">Client</Link>
+      <Link href="/pump">Pump</Link>
+    </div>
   );
 }